+++ src/extensions.ts
@@ -19,13 +19,13 @@ import * as stream from 'node:stream'
 import * as errors from './errors.ts'
 import { isBoolean, isNumber, isString, isValidBucketName, isValidPrefix, pipesetup, uriEscape } from './helpers.ts'
 import * as transformers from './transformers.ts'
-import type { TypedClient2 } from './typed-client2.ts'
+import type { Client } from './typed-client2.ts'
 
 // TODO
 type S3Object = unknown
 
 export class extensions {
-  constructor(readonly client: TypedClient2) {}
+  constructor(readonly client: Client) {}
 
   // List the objects in the bucket using S3 ListObjects V2 With Metadata
   //
